refactor: separate rootReducer to separate individual file
